fix paddle integration test fixture import
